+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # 0.1.0
+## 2016-09-27
+* Added a `purgeEverything` function to CloudflarePurger.
+
+
 ## 2016-09-26
 * Add logic to Purge draft of pages.